requirements.txt
aicsimageio==4.7.0
aicssegmentation==0.5.3
alabaster==0.7.13
annotated-types==0.5.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asciitree==0.3.3
attrs==23.2.0
aws-requests-auth==0.4.3
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.0.0
boto3==1.33.13
botocore==1.33.13
cachey==0.2.1
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
colorcet==3.1.0
comm==0.1.4
cycler==0.11.0
dask==2022.2.0
debugpy==1.7.0
decorator==5.1.1
defusedxml==0.7.1
docstring_parser==0.16
docutils==0.19
elementpath==4.1.5
entrypoints==0.4
fasteners==0.19
fastjsonschema==2.19.1
fonttools==4.38.0
freetype-py==2.4.0
fsspec==2023.1.0
future==1.0.0
graphviz==0.20.1
HeapDict==1.0.1
hsluv==5.0.4
idna==3.6
imagecodecs==2021.11.20
imageio==2.31.2
imagesize==1.4.1
imbalanced-learn==0.11.0
importlib-metadata==6.7.0
importlib-resources==5.12.0
ipydatawidgets==4.3.2
ipykernel==6.16.2
ipympl==0.9.3
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==7.8.1
itk==5.3.0
itk-core==5.3.0
itk-filtering==5.3.0
itk-io==5.3.0
itk-meshtopolydata==0.10.0
itk-numerics==5.3.0
itk-registration==5.3.0
itk-segmentation==5.3.0
itkwidgets==0.32.6
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonlines==1.2.0
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.7
kiwisolver==1.4.5
locket==1.0.0
lxml==4.9.4
magicgui==0.6.1
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==3.0.2
mypy-extensions==1.0.0
napari==0.4.14
napari-console==0.0.7
napari-plugin-engine==0.2.0
napari-svg==0.1.10
nbclient==0.7.4
nbconvert==7.6.0
nbformat==5.8.0
nest-asyncio==1.6.0
networkx==2.6.3
notebook==6.4.13
npe2==0.2.1
numcodecs==0.10.2
numpy==1.21.6
numpydoc==1.5.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
ome-types==0.5.0
opencv-python==4.9.0.80
packaging==24.0
pandas==1.3.5
pandocfilters==1.5.1
parso==0.8.3
partd==1.4.1
patsy==0.5.6
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
Pint==0.18
pkgutil_resolve_name==1.3.10
platformdirs==4.0.0
prometheus-client==0.17.1
prompt-toolkit==3.0.43
psutil==5.9.8
psygnal==0.9.5
ptyprocess==0.7.0
pyarrow==12.0.1
pycparser==2.21
pydantic==2.5.3
pydantic-compat==0.1.2
pydantic_core==2.14.6
Pygments==2.17.2
PyOpenGL==3.1.7
pyparsing==3.1.2
pyrsistent==0.19.3
python-dateutil==2.9.0.post0
pytomlpp==1.0.13
pytz==2024.1
PyWavelets==1.3.0
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.4.4
QtPy==2.4.1
quilt==2.9.15
quilt3==5.4.0
requests==2.31.0
requests-futures==1.0.0
s3transfer==0.8.2
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
Send2Trash==1.8.2
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.4.1
Sphinx==5.3.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
statsmodels==0.13.5
superqt==0.5.0
tenacity==8.2.3
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2021.11.2
tinycss2==1.2.1
toolz==0.12.1
torch==1.13.1
torchmetrics==0.11.4
torchview==0.2.6
torchvision==0.14.1
torchviz==0.0.2
tornado==6.2
tqdm==4.66.2
traitlets==5.6.0
traittypes==0.2.1
typer==0.9.0
typing_extensions==4.7.1
urllib3==1.26.18
vispy==0.13.0
wcwidth==0.2.13
webencodings==0.5.1
widgetsnbextension==3.6.6
wrapt==1.16.0
xarray==0.20.2
xlrd==2.0.1
xmlschema==2.5.1
xsdata==23.7
zarr==2.12.0
zipp==3.15.0
zstandard==0.21.0
